Exporting extractor list for use in other clients
